creatinga async queue for topper filter
