+++ cron/tasks.py
@@ -21,7 +21,7 @@ from django.core.exceptions import ValidationError
 from smtplib import SMTPException, SMTPServerDisconnected
 from django.core.mail import BadHeaderError
 from rq.decorators import job
-from cron import REDIS_CLIENT, DEFAULT_QUEUE
+from cron import REDIS_CLIENT, DEFAULT_QUEUE, TOPPER_QUEUE
 from rq import Retry
 import time
 from rq import get_current_job
@@ -108,4 +108,46 @@ def bulk_email(taskid, *args, **kwargs):
 
 def async_bulk_email(task, *args, **kwargs):
     DEFAULT_QUEUE.enqueue(bulk_email, task.pk, job_id=task.job_id, job_timeout='72h')
-    print('working')
+
+
+def filter_student_grades(foss=None, state=None, city=None, grade=None, institution_type=None, activation_status=None, from_date=None, to_date=None, key=None):
+  if grade:
+    #get the moodle id for the foss
+    try:
+      fossmdl=FossMdlCourses.objects.filter(foss__in=foss)
+      #get moodle user grade for a specific foss quiz id having certain grade
+      user_grade=MdlQuizGrades.objects.using('moodle').values_list('userid', 'quiz', 'grade').filter(quiz__in=[f.mdlquiz_id for f in fossmdl], grade__gte=int(grade))
+      #convert moodle user and grades as key value pairs
+      dictgrade = {i[0]:{i[1]:[i[2],False]} for i in user_grade}
+      #get all test attendance for moodle user ids and for a specific moodle quiz ids
+      test_attendance=TestAttendance.objects.filter(
+                                mdluser_id__in=list(dictgrade.keys()), 
+                                mdlquiz_id__in=[f.mdlquiz_id for f in fossmdl], 
+                                test__academic__state__in=state if state else State.objects.all(),
+                                test__academic__city__in=city if city else City.objects.all(), 
+                                status__gte=3, 
+                                test__academic__institution_type__in=institution_type if institution_type else InstituteType.objects.all(), 
+                                test__academic__status__in=[activation_status] if activation_status else [1,3]
+                              )
+      if from_date and to_date:
+        test_attendance = test_attendance.filter(test__tdate__range=[from_date, to_date])
+      elif from_date:
+        test_attendance = test_attendance.filter(test__tdate__gte=from_date)
+        
+      filter_ta=[]
+      for i in range(test_attendance.count()):
+        if not dictgrade[test_attendance[i].mdluser_id][test_attendance[i].mdlquiz_id][1]:
+          dictgrade[test_attendance[i].mdluser_id][test_attendance[i].mdlquiz_id][1] = True
+          filter_ta.append(test_attendance[i])
+          
+      #return the result as dict
+      result= {'mdl_user_grade': dictgrade, 'test_attendance': filter_ta, "count":len(filter_ta)}
+      caches['file_cache'].set(key,result)
+      return
+    except FossMdlCourses.DoesNotExist:
+      return None
+  return None
+
+
+def async_filter_student_grades(foss=None, state=None, city=None, grade=None, institution_type=None, activation_status=None, from_date=None, to_date=None, key=None):
+    TOPPER_QUEUE.enqueue(filter_student_grades, (foss, state, city, grade, institution_type, activation_status, from_date, to_date, key), job_id=key, job_timeout='72h')
